requirements.txt
appdirs==1.4.4
asgiref==3.4.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.2.6
django-3-jet==1.0.8
django-bootstrap-form==3.4
django-debug-toolbar==3.2.2
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.13.2
djangorestframework==3.13.1
docopt==0.6.2
idna==2.10
importlib-metadata==4.11.3
Markdown==3.3.7
mypy-extensions==0.4.3
nose==1.3.7
num2words==0.5.10
numpy==1.21.2
pathspec==0.8.1
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
stripe==2.56.0
toml==0.10.2
tornado==6.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
xlwt==1.3.0
zipp==3.8.0
